requirements.txt
# Core Dependencies
PySide6>=6.6.0
pycryptodome>=3.19.0
python-dateutil>=2.8.2
